# openapi.yaml
openapi: 3.0.0
info:
  title: book-comet
  version: '1.0'
servers:
  - url: 'http://localhost:8000'
    description: Local
paths:
...
  '/inventory/{bookId}/increase':
    parameters:
      - schema:
          type: string
        name: bookId
        in: path
        required: true
        description: Book id
    put:
      summary: Book Inventory Increase
      operationId: inventory-book-increase
      responses:
        '204':
          description: No Content
      description: Increase book inventory amount by value passed in the body
      requestBody:
        $ref: '#/components/requestBodies/UpdateAmount'
      parameters:
        - $ref: '#/components/parameters/Session'
...
  '/inventory/{bookId}/decrease':
    parameters:
      - schema:
          type: string
        name: bookId
        in: path
        required: true
        description: Book id
    put:
      summary: Book Inventory Decrease
      tags: []
      responses:
        '204':
          description: No Content
      operationId: get-inventory-bookId-decrease
      description: Decrease book inventory amount by value passed in the body
      requestBody:
        $ref: '#/components/requestBodies/UpdateAmount'
      parameters:
        - $ref: '#/components/parameters/Session'
...
components:
  schemas:
...
    UpdateAmountModel:
      title: IncreaseAmount
      type: object
      properties:
        amount:
          type: integer
          minimum: 1
      required:
        - amount
...
  securitySchemes: {}
  parameters:
    Session:
      name: connect.sid
      in: cookie
      required: true
      schema:
        type: string
      description: Login Session Cookie
...
  requestBodies:
...
    UpdateAmount:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdateAmountModel'